About Java-Project-on-Laptop-Catalogue
Developing a standalone project using JAVA and for database connectivity SQL Server 2008 on netbens.
My Project is based upon the management of the products available in the company. 
It is a commercially based project.
It manages the various types of products that are available and make it easy for the customers to order there desire product.